[Optimization-2703][studio] Optimize online debug task
